implement getFolder and magically delete the folders up to the module root in case the parent folders no longer contain files

+++ src/fkooman/RemoteStorage/Document.php
@@ -53,12 +53,84 @@ class Document
 
     public function deleteDocument(Path $p)
     {
+        if ($p->getIsFolder()) {
+            throw new DocumentException("unable to delete folder");
+        }
+
         $documentPath = $this->baseDir . $p->getPath();
+
         if (false === file_exists($documentPath)) {
             throw new DocumentMissingException();
         }
+
         if (false === @unlink($documentPath)) {
             throw new DocumentException("unable to delete file");
+        }
+
+        // delete all empty folders in the tree up to the module root if
+        // they are empty
+        $p = new Path($p->getParentFolder());
+        while (!$p->getIsModuleRoot()) {
+            // not the module root
+            if ($this->isEmptyFolder($p)) {
+                // and it is empty, delete it
+                $this->deleteFolder($p);
+            }
+            $p = new Path($p->getParentFolder());
+        }
+    }
+
+    public function getFolder(Path $p)
+    {
+        if (!$p->getIsFolder()) {
+            throw new DocumentException("not a folder");
+        }
+
+        $folderPath = $this->baseDir . $p->getPath();
+
+        $entries = glob($folderPath . "*", GLOB_ERR|GLOB_MARK);
+        if (false === $entries) {
+            // directory does not exist, return empty list
+            return array();
+        }
+        $folderEntries = array();
+        foreach ($entries as $e) {
+            if (is_dir($e)) {
+                $folderEntries[basename($e) . "/"] = array();
+            } else {
+                $folderEntries[basename($e)] = array(
+                    "Content-Length" => filesize($e)
+                );
+            }
+        }
+
+        return $folderEntries;
+    }
+
+    private function isEmptyFolder(Path $p)
+    {
+        if (!$p->getIsFolder()) {
+            throw new DocumentException("not a folder");
+        }
+
+        $folderPath = $this->baseDir . $p->getPath();
+
+        $entries = glob($folderPath . "*", GLOB_ERR);
+        if (false === $entries) {
+            throw new DocumentException("unable to read folder");
+        }
+
+        return 0 === count($entries);
+    }
+
+    private function deleteFolder(Path $p)
+    {
+        if (!$p->getIsFolder()) {
+            throw new DocumentException("not a folder");
+        }
+        $folderPath = $this->baseDir . $p->getPath();
+        if (false === @rmdir($folderPath)) {
+            throw new DocumentException("unable to delete folder");
         }
     }
 }
